requirements.txt
Django==5.1.4
channels==4.2.0
channels-redis==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
daphne==4.2.0
django-cors-headers==4.3.1
PyJWT==2.8.0